requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
billiard==3.6.3.0
boto3==1.12.39
botocore==1.15.39
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
daphne==2.4.1
defusedxml==0.6.0
dj-static==0.0.6
Django==2.2
django-admin-rangefilter==0.5.4
django-admin-views==0.8.0
django-celery-beat==1.1.1
django-celery-results==1.1.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-extra-fields==2.0.4
django-map-widgets==0.3.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-rest-recaptcha==1.0.2
django-rest-swagger==2.2.0
django-sendgrid-v5==0.8.1
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils<0.16
future==0.18.2
googlemaps==4.2.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
itypes==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
kombu==4.6.8
MarkupSafe==1.1.1
msgpack~=0.6.0
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.1
postgis==1.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pystache==0.5.4
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.2.1
sentry-sdk==0.14.3
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
static==1.1.1
static3==0.7.0
twilio==6.38.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
django-leaflet==0.26.0
django-rest-knox==4.1.0
django-twilio==0.11.0
phonenumbers==8.11.1